requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# LangGraph & LangChain
langgraph==0.2.16
langchain>=0.1.0
langchain-core>=0.2.27,<0.3

# AI Models
anthropic==0.8.1
openai==1.6.1
google-generativeai==0.3.2

# Automation Drivers
playwright==1.40.0
appium-python-client==3.1.0
selenium==4.15.0

# Document Processing
PyMuPDF==1.23.14
pytesseract==0.3.10
Pillow==10.1.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.0
typing-extensions==4.8.0
asyncio==3.4.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
